Move key generate rewrite logic to infra module for reuse
