+++ infra/rewrite/src/main/java/org/apache/shardingsphere/infra/rewrite/token/generator/keygen/BaseGeneratedKeyTokenGenerator.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.sharding.rewrite.token.generator.impl.keygen;
+package org.apache.shardingsphere.infra.rewrite.token.generator.keygen;
 
 import org.apache.shardingsphere.infra.rewrite.sql.token.generator.OptionalSQLTokenGenerator;
 import org.apache.shardingsphere.infra.binder.context.statement.SQLStatementContext;
