+++ infra/rewrite/src/main/java/org/apache/shardingsphere/infra/rewrite/token/generator/keygen/GeneratedKeyInsertColumnTokenGenerator.java
@@ -15,10 +15,10 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.sharding.rewrite.token.generator.impl.keygen;
+package org.apache.shardingsphere.infra.rewrite.token.generator.keygen;
 
 import com.google.common.base.Preconditions;
-import org.apache.shardingsphere.sharding.rewrite.token.pojo.GeneratedKeyInsertColumnToken;
+import org.apache.shardingsphere.infra.rewrite.token.pojo.keygen.GeneratedKeyInsertColumnToken;
 import org.apache.shardingsphere.infra.binder.context.segment.insert.keygen.GeneratedKeyContext;
 import org.apache.shardingsphere.infra.binder.context.statement.dml.InsertStatementContext;
 import org.apache.shardingsphere.sql.parser.sql.common.segment.dml.column.InsertColumnsSegment;
